+++ src/GravitySimulation_v2/GravitySimulation_v2/globals/globals.cpp
@@ -1,5 +1,5 @@
 #include "globals.h"
-#include "grid.h"
+#include "../grid/grid.h"
 
 std::vector<Particle> particles;
