+++ src/GravitySimulation_v2/GravitySimulation_v2/globals/setup_type.h
@@ -1,8 +1,6 @@
 #ifndef SETUPTYPE_H
 #define SETUPTYPE_H
 
-#include "settings.h"
-
 enum class PlacementType
 {
     Fullscreen,
